requirements.txt
Django==4.0.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0

PyJWT==2.3.0
pytz==2022.1
mysqlclient==2.1.0
